requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
colorama==0.4.6
importlib_resources==6.5.2
pillow==11.3.0
pygame==2.6.1
pyserial==3.5
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
pywin32==311
PyYAML==6.0.3
qrcode==8.2
setuptools==80.9.0
six==1.17.0
